ohos_deviceinfo_sys

Module deviceinfo_ffi

Functions

OH_GetAbiList⚠

Obtains the application binary interface (Abi) list represented by a string.

OH_GetBootloaderVersion⚠

Obtains the bootloader version number represented by a string.

OH_GetBrand⚠

Obtains the device brand represented by a string.

OH_GetBuildHost⚠

Obtains the build host of the current running OS.

OH_GetBuildRootHash⚠

Obtains the version hash of the current running OS.

OH_GetBuildTime⚠

Obtains the build time of the current running OS.

OH_GetBuildType⚠

Obtains the build type of the current running OS.

OH_GetBuildUser⚠

Obtains the build user of the current running OS.

OH_GetDeviceType⚠

Obtains the device type represented by a string, which can be phone (or default for phones), wearable, liteWearable, tablet, tv, car, or smartVision.

OH_GetDisplayVersion⚠

Obtains the product version displayed for customer represented by a string.

OH_GetDistributionOSApiVersion⚠

Obtains the ISV distribution OS api version represented by a integer. If ISV not specified, it will return the same value as OH_GetSdkApiVersion

OH_GetDistributionOSName⚠

Obtains the Distribution OS name represented by a string.

OH_GetDistributionOSReleaseType⚠

Obtains the ISV distribution OS release type represented by a string. If ISV not specified, it will return the same value as OH_GetOsReleaseType

OH_GetDistributionOSVersion⚠

Obtains the ISV distribution OS version represented by a string. If ISV not specified, it will return the same value as OH_GetOSFullName

OH_GetFirstApiVersion⚠

Obtains the first API version number.

OH_GetHardwareModel⚠

Obtains the hardware model represented by a string.

OH_GetIncrementalVersion⚠

Obtains the incremental version represented by a string.

OH_GetManufacture⚠

Obtains the device manufacturer represented by a string.

OH_GetMarketName⚠

Obtains the product name speaded in the market

OH_GetOSFullName⚠

Obtains the OS full version name represented by a string.

OH_GetOsReleaseType⚠

Obtains the OS release type represented by a string.

OH_GetProductModel⚠

Obtains the product model represented by a string.

OH_GetProductSeries⚠

Obtains the product series represented by a string.

OH_GetSdkApiVersion⚠

Obtains the SDK API version number.

OH_GetSecurityPatchTag⚠

Obtains the security patch tag represented by a string.

OH_GetSoftwareModel⚠

Obtains the software model represented by a string.

OH_GetVersionId⚠

Obtains the version ID by a string.